+++ __docs__/wonder-blocks-badge/badge-testing-snapshots.stories.tsx
@@ -362,10 +362,15 @@ export const AllBadgesScenarios: StoryComponentType = {
                                         backgroundColor:
                                             semanticColor.surface.inverse,
                                         borderColor:
-                                            semanticColor.border.inverse,
-                                        color: semanticColor.text.inverse,
+                                            semanticColor.core.border.inverse
+                                                .strong,
+                                        color: semanticColor.core.foreground
+                                            .inverse.strong,
                                     },
-                                    icon: {color: semanticColor.icon.inverse},
+                                    icon: {
+                                        color: semanticColor.core.foreground
+                                            .inverse.strong,
+                                    },
                                     label: {
                                         fontWeight: font.weight.medium,
                                     },