+++ __docs__/wonder-blocks-badge/badge.stories.tsx
@@ -201,10 +201,10 @@ export const CustomStyles: StoryComponentType = {
                 styles={{
                     root: {
                         backgroundColor: semanticColor.surface.inverse,
-                        borderColor: semanticColor.border.inverse,
-                        color: semanticColor.text.inverse,
+                        borderColor: semanticColor.core.border.inverse.strong,
+                        color: semanticColor.core.foreground.inverse.strong,
                     },
-                    icon: {color: semanticColor.icon.inverse},
+                    icon: {color: semanticColor.core.foreground.inverse.strong},
                     label: {
                         fontWeight: font.weight.medium,
                     },