+++ __docs__/wonder-blocks-icon/custom-icon-components.stories.tsx
@@ -110,7 +110,7 @@ export const CustomIconsWithCustomStyle: StoryComponentType = {
             backgroundColor: semanticColor.surface.secondary,
             padding: sizing.size_040,
             borderRadius: border.radius.radius_040,
-            border: `${border.width.thin} solid ${semanticColor.border.subtle}`,
+            border: `${border.width.thin} solid ${semanticColor.core.border.neutral.subtle}`,
         },
     },
 };