+++ packages/wonder-blocks-modal/src/components/__tests__/focus-trap.test.tsx
@@ -8,7 +8,7 @@ import {Choice, RadioGroup} from "@khanacademy/wonder-blocks-form";
 import FocusTrap from "../focus-trap";
 
 describe("FocusTrap", () => {
-    it("Focus should move to the first focusable element", async () => {
+    it("moves focus to the first focusable element", async () => {
         // Arrange
         render(
             <>
@@ -37,7 +37,7 @@ describe("FocusTrap", () => {
         );
 
         // Initial focused element
-        const firstRadioButton = await screen.findByRole("radio", {
+        const firstRadioButton = screen.getByRole("radio", {
             name: /first option/i,
         });
         firstRadioButton.focus();
@@ -53,7 +53,7 @@ describe("FocusTrap", () => {
         expect(firstRadioButton).toHaveFocus();
     });
 
-    it("Focus should move to the last focusable element", async () => {
+    it("moves focus to the last focusable element", async () => {
         // Arrange
         render(
             <>
@@ -82,7 +82,7 @@ describe("FocusTrap", () => {
         );
 
         // Initial focused element
-        const firstRadioButton = await screen.findByRole("radio", {
+        const firstRadioButton = screen.getByRole("radio", {
             name: /first option/i,
         });
         firstRadioButton.focus();
@@ -94,7 +94,51 @@ describe("FocusTrap", () => {
         // Assert
         // Redirect focus to the button.
         expect(
-            await screen.findByRole("button", {name: "A focusable button"}),
+            screen.getByRole("button", {name: "A focusable button"}),
         ).toHaveFocus();
+    });
+
+    it("does not move focus to elements with display: none", async () => {
+        // Arrange
+        render(
+            <FocusTrap>
+                <Button>button 1</Button>
+                <Button>button 2</Button>
+                <Button id="hidden-element" style={{display: "none"}}>
+                    button 3
+                </Button>
+            </FocusTrap>,
+        );
+
+        // Initial focused element
+        screen.getByRole("button", {name: /button 1/i}).focus();
+
+        // Act
+        await userEvent.tab({shift: true});
+
+        // Assert
+        expect(screen.getByRole("button", {name: /button 2/i})).toHaveFocus();
+    });
+
+    it("does not move focus to elements with visibility: hidden", async () => {
+        // Arrange
+        render(
+            <FocusTrap>
+                <Button>button 1</Button>
+                <Button>button 2</Button>
+                <Button id="hidden-element" style={{visibility: "hidden"}}>
+                    button 3
+                </Button>
+            </FocusTrap>,
+        );
+
+        // Initial focused element
+        screen.getByRole("button", {name: /button 1/i}).focus();
+
+        // Act
+        await userEvent.tab({shift: true});
+
+        // Assert
+        expect(screen.getByRole("button", {name: /button 2/i})).toHaveFocus();
     });
 });