+++ packages/wonder-blocks-toolbar/src/components/toolbar.tsx
@@ -146,7 +146,7 @@ const sharedStyles = StyleSheet.create({
     dark: {
         background: semanticColor.surface.inverse,
         boxShadow: `0 1px 0 0 ${color.white64}`,
-        color: semanticColor.text.inverse,
+        color: semanticColor.core.foreground.inverse.strong,
     },
     leftColumn: {
         alignItems: "center",
@@ -160,7 +160,7 @@ const sharedStyles = StyleSheet.create({
         flexGrow: 1,
     },
     subtitle: {
-        color: semanticColor.text.secondary,
+        color: semanticColor.core.foreground.neutral.default,
     },
     titles: {
         padding: spacing.small_12,