+++ packages/wonder-blocks-icon/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @khanacademy/wonder-blocks-icon
 
+## 5.2.12
+
+### Patch Changes
+
+- Updated dependencies [eafcfa3]
+- Updated dependencies [956f8bd]
+- Updated dependencies [eafcfa3]
+- Updated dependencies [b42b694]
+- Updated dependencies [956f8bd]
+    - @khanacademy/wonder-blocks-tokens@12.0.0
+
+## 5.2.11
+
+### Patch Changes
+
+- Updated dependencies [03415d1]
+    - @khanacademy/wonder-blocks-tokens@11.4.1
+
 ## 5.2.10
 
 ### Patch Changes