+++ packages/wonder-blocks-search-field/CHANGELOG.md
@@ -1,5 +1,41 @@
 # @khanacademy/wonder-blocks-search-field
 
+## 5.1.36
+
+### Patch Changes
+
+- Updated dependencies [e934f20]
+    - @khanacademy/wonder-blocks-form@7.3.0
+
+## 5.1.35
+
+### Patch Changes
+
+- eafcfa3: Replaces `semanticColor.icon` with `semanticColor.core.foreground` tokens
+- Updated dependencies [956f8bd]
+- Updated dependencies [eafcfa3]
+- Updated dependencies [956f8bd]
+- Updated dependencies [956f8bd]
+- Updated dependencies [eafcfa3]
+- Updated dependencies [b42b694]
+- Updated dependencies [956f8bd]
+    - @khanacademy/wonder-blocks-form@7.2.5
+    - @khanacademy/wonder-blocks-tokens@12.0.0
+    - @khanacademy/wonder-blocks-icon@5.2.12
+    - @khanacademy/wonder-blocks-icon-button@10.3.13
+    - @khanacademy/wonder-blocks-typography@4.2.14
+
+## 5.1.34
+
+### Patch Changes
+
+- Updated dependencies [03415d1]
+    - @khanacademy/wonder-blocks-tokens@11.4.1
+    - @khanacademy/wonder-blocks-form@7.2.4
+    - @khanacademy/wonder-blocks-icon@5.2.11
+    - @khanacademy/wonder-blocks-icon-button@10.3.12
+    - @khanacademy/wonder-blocks-typography@4.2.13
+
 ## 5.1.33
 
 ### Patch Changes