+++ packages/wonder-blocks-styles/CHANGELOG.md
@@ -1,5 +1,24 @@
 # @khanacademy/wonder-blocks-styles
 
+## 0.2.25
+
+### Patch Changes
+
+- 956f8bd: Replaces instances of `semanticColor.text` with `semanticColor.core.foreground`
+- Updated dependencies [eafcfa3]
+- Updated dependencies [956f8bd]
+- Updated dependencies [eafcfa3]
+- Updated dependencies [b42b694]
+- Updated dependencies [956f8bd]
+    - @khanacademy/wonder-blocks-tokens@12.0.0
+
+## 0.2.24
+
+### Patch Changes
+
+- Updated dependencies [03415d1]
+    - @khanacademy/wonder-blocks-tokens@11.4.1
+
 ## 0.2.23
 
 ### Patch Changes