+++ packages/wonder-blocks-toolbar/CHANGELOG.md
@@ -1,5 +1,26 @@
 # @khanacademy/wonder-blocks-toolbar
 
+## 5.1.31
+
+### Patch Changes
+
+- 956f8bd: Replaces instances of `semanticColor.text` with `semanticColor.core.foreground`
+- Updated dependencies [eafcfa3]
+- Updated dependencies [956f8bd]
+- Updated dependencies [eafcfa3]
+- Updated dependencies [b42b694]
+- Updated dependencies [956f8bd]
+    - @khanacademy/wonder-blocks-tokens@12.0.0
+    - @khanacademy/wonder-blocks-typography@4.2.14
+
+## 5.1.30
+
+### Patch Changes
+
+- Updated dependencies [03415d1]
+    - @khanacademy/wonder-blocks-tokens@11.4.1
+    - @khanacademy/wonder-blocks-typography@4.2.13
+
 ## 5.1.29
 
 ### Patch Changes